account_service/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
python-dotenv
python-jose[cryptography]
 pydantic-settings
python-multipart